% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{cal_betas}
\alias{cal_betas}
\title{calculates betacoefficients from explained variances}
\usage{
cal_betas(pves, maf, gy_direct = T)
}
\arguments{
\item{pves}{a named list of parameters and percentages of variance explained for each one.
If multiple snps, each param is a list of snp-vectors, each vector the scenarios for this snp}

\item{maf}{a named list of maf for every snp}

\item{gy_direct}{should there be a direct effect of G on Y. default true. does not need to be changed.}
}
\value{
a list of two data.tables. a data.table containing all valid combinations
    of coefficients, one row is one snp in one scenario. A data.table with the adjusted combinations
    of percentages of variance explained
}
\description{
For examples see vignette browseVignettes('MRTool')
}
