% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulating.R
\name{CreateMRInputObject}
\alias{CreateMRInputObject}
\title{Create MRInput Object}
\usage{
CreateMRInputObject(MR_data = my_MR_data, MR_Scenario = "V1", correlation = T)
}
\arguments{
\item{correlation}{}
}
\description{
For examples see vignette browseVignettes('MRTool')
}
