% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{get_corrs}
\alias{get_corrs}
\title{calculate correlation for a datatable of snps}
\usage{
get_corrs(genedose_test, method = "pearson")
}
\arguments{
\item{genedose_test}{snp datatable. first two columns identifiers}

\item{method}{used to compute correlation coefficient. 'pearson' (default), 'kendall' or 'spearman'}
}
\description{
returns 2 snps each that have lowest,
positive high, positive low, negative high, negative low correlation. uses pearson correlation
For examples see vignette browseVignettes('MRTool')
}
