% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{SetMRParams}
\alias{SetMRParams}
\title{Set MR Parameters (betacoefficients)}
\usage{
SetMRParams(
  Vary_between_SNP = F,
  G_X_randomization = F,
  SNP = NULL,
  G_U = c(0, 0.1),
  G_Y = c(0, 0.1),
  X_Y = c(0, 0.1),
  G_X = c(0.03, 0.1),
  U_X = c(0, 0.1),
  U_Y = c(0, 0.1),
  sim_sd = 1,
  maf
)
}
\arguments{
\item{maf}{}
}
\description{
Create a testplan for different parameter combinations
or create a custom plan for different IV parameters. If Vary_between_SNP, The Parameters set
for G_U, G_X and G_Y will not create individual scenarios but one, where the different
parameters are distributed to the individual SNP, cycled through all give SNP, based on the
number of parameters provided by the user, accepting only one parameter (the first)
for U_Y, U_X and X_Y
For examples see vignette browseVignettes('MRTool')
}
