% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulating.R
\name{sim_and_mr}
\alias{sim_and_mr}
\title{Simulate U,X and Y and run MR on them}
\usage{
sim_and_mr(SNP_data, SNP, my_Parameters, reverse = FALSE)
}
\arguments{
\item{reverse=FALSE}{TRUE if MR should be conducted with X as putative outcome
and Y as putative exposure}
}
\value{
list(reg_res,sim_data,mr_res)
}
\description{
Simulates confounder U, exposure X and outcome Y based on genotype data and betacoefficients.
For examples see vignette browseVignettes('MRTool')
}
