% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulating.R
\name{SimulateMRData}
\alias{SimulateMRData}
\title{Simulate MR Data}
\usage{
SimulateMRData(SNP_data = SNP_data, SNP = NULL, Parameters = NULL)
}
\arguments{
\item{Parameters}{}
}
\description{
For examples see vignette browseVignettes('MRTool')
}
