% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{transform_results}
\alias{transform_results}
\title{Transform result of iterations of simulation and MR}
\usage{
transform_results(results, SNP, iterations, pve_grid, rev = FALSE)
}
\arguments{
\item{results}{list of results of single iteration. result of single iteration is list(reg_res,sim_data,mr_res)}

\item{SNP}{names of snps used}

\item{iterations}{integer how many times the simulation was repeated}

\item{pve_grid}{grid of percentages of variance explained as output of cal_betas}

\item{rev}{true if mr was executed in the anticausal direction. false for causal direction}
}
\value{
list(res_list,sim_data). res_list
is list of regression results and mr results. sim_data is result of simulation
}
\description{
Transforms results of x iterations with each n scenario simulations into median per scenario
for relevant variables.
For examples see vignette browseVignettes('MRTool')
}
